+++ server/controllers/verifyTokenController.js
@@ -1,5 +1,15 @@
 const jwt = require('jsonwebtoken');
 
 const tokenVerifier = (req, res, next) => {
+    const bearerHeader = req.headers['authorization']
+    if(typeof bearerHeader !== 'undefined') {
+      const bearer = bearerHeader.split(' ')
+      const bearerToken = bearer[1]
+      req.token = bearerToken
+      next()
+    } else {
+      res.sendStatus(403)
+    }
+  }
 
-}
+  module.exports = tokenVerifier;
